requirements.txt
openai
langchain
python-dotenv
wikidata
wikipedia
